+++ src/hooks.ts
@@ -1,7 +1,9 @@
 import {
     CreaturePF2e,
     CreatureSheetPF2e,
+    EncounterPF2e,
     ItemSheetPF2e,
+    NPCPF2e,
     NPCSheetPF2e,
     PhysicalItemPF2e,
 } from "foundry-pf2e";
@@ -12,6 +14,7 @@ import { RefinedItem } from "./refined-item";
 import { configureRefinedItem } from "./app/refined-item-editor";
 import { createRefinedItemDialog } from "./app/refined-item-create";
 import { MODULE_ID } from "./module";
+import { AfterCombatDialog } from "./app/after-combat-dialog";
 
 export class ModuleHooks {
     static registerAllHandlers() {
@@ -23,6 +26,13 @@ export class ModuleHooks {
             "none"
         )
             this.hideSellMonsterPartsButton();
+
+        if (
+            game.user.isActiveGM &&
+            game.settings.get(MODULE_ID, "monster-parts-after-combat") !==
+                "none"
+        )
+            this.registerEndCombatDialog();
     }
 
     static addMonsterPartButton() {
@@ -150,5 +160,45 @@ export class ModuleHooks {
                 });
             },
         );
+    }
+
+    static registerEndCombatDialog() {
+        Hooks.on("deleteCombat", async (combat: EncounterPF2e) => {
+            const combatants = combat.combatants;
+            const npcs = combatants
+                .filter((c) => (c.token && c.actor?.isOfType("npc")) ?? false)
+                .map((e) => ({
+                    token: e.token!,
+                    actor: e.actor! as NPCPF2e,
+                    value: MonsterPart.valueOfCreature(e.actor as NPCPF2e),
+                    checked: e.actor!.isDead,
+                }));
+            if (npcs.length == 0) return;
+            const settings = game.settings.get(
+                MODULE_ID,
+                "monster-parts-after-combat",
+            );
+            const dialogResult =
+                settings == "dialog"
+                    ? ((await new Promise((resolve) => {
+                          const d = new AfterCombatDialog({
+                              npcs,
+                              form: {
+                                  handler: async (
+                                      _event: Event | SubmitEvent,
+                                      _form: HTMLFormElement,
+                                      _formData: foundry.applications.ux.FormDataExtended,
+                                  ) => resolve(d.npcs),
+                              },
+                          });
+                          d.render(true);
+                      })) as typeof npcs | undefined)
+                    : npcs;
+
+            if (!dialogResult) return;
+            dialogResult
+                .filter((e) => e.checked)
+                .forEach((e) => MonsterPart.fromCreature(e.actor));
+        });
     }
 }